Add drive detail, assignments, call center, and live dashboard components
- Introduced `DriveDetailContent` component for displaying detailed drive information.
- Created `AssignmentsContent` component for managing volunteer assignments with drag-and-drop functionality.
- Implemented `CallCenterContent` for managing volunteer calls and tracking call results.
- Added `LiveDashboardContent` to provide real-time updates on volunteer statuses during drives.
- Refactored existing pages to utilize new components, enhancing code organization and maintainability.
